pyproject.toml
[project]
name = "kyc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "pytest>=8.4.2",
]
